requirements.txt
python-telegram-bot==13.7
psycopg2==2.9.1
